chatgpt_assistant/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
flask==2.3.3
apscheduler==3.10.1
openai==1.3.0
google-api-python-client==2.108.0
beautifulsoup4==4.12.2
pandas==2.1.0
pyyaml==6.0.1